サムネイル・タイトル取得のメソッド定義場所を修正
